ForwardBackward

Documentation for ForwardBackward.

ForwardBackward.Deterministic

ForwardBackward.ManifoldProcess

ForwardBackward.perturb!

ForwardBackward.stochastic

ForwardBackward.Deterministic — Type

Deterministic()

Endpoint conditioning under a deterministic process is a linear interpolation.

ForwardBackward.ManifoldProcess — Type

ManifoldProcess(v::T)

A process on a manifold, with a drift variance v.

ForwardBackward.perturb! — Method

perturb(M, p, v)

Perturb a point on a manifold by a normal distribution with variance v.

ForwardBackward.stochastic — Method

stochastic(o::State)
stochastic(T::Type, o::State)

Converts a state to a distribution. Default type is Float64.